chore: expose ack api
