+++ src/queries/mod.rs
@@ -2,6 +2,12 @@
 //!
 //! Each submodule contains specific query logic used by the backend.
 use apalis_core::backend::StatType;
+
+/// ACK a task on success or failure
+pub mod ack_task;
+
+/// Lock a task for processing
+pub mod lock_task;
 /// Fetch tasks by their IDs
 pub mod fetch_by_id;
 /// Keep workers alive by updating their heartbeat
